requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.8.1
django-multiselectfield==0.1.12
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-tempus-dominus==5.1.2.13
html5lib==1.0.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.20
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
